Link food and recipe names to dontstarve.wikia.com
- Added wikiaHref() function, which returns an <a href=...> document
element, from a food/recipe name.

- Both makeFoodRow() and makeRecipeRow() use wikiaHref() instead
of the plain item.name string.

- cells() needed an additional if-clause to handle elements, otherwise
the html markup will get escaped, and the food/recipe names will show up
as something like '<a href="...">Banana</a>'.

- One edge case for the string "Combined" that should *not* get linked
to wikia.com, and so wikiaHref() simply returns that unchanged, rather
than creating a document element for it.