Hook, Line, & Inker (Syntax Error Correction)
Thanks to Hornette on the Klei forums for providing the images for the new fish.

Changes/Additions:
Kelp, Cooked Kelp, and Dried Kelp are now allowed in the crockpot (0.5 Veggie). Note: Dried Kelp is currently bugged in-game where it can't be placed in the crockpot. Should be fixed by Klei within a few days.
10 New fishes that can be put in the crockpot.
Fish Morsels and Fish Meat are a part of DST now (only Fish Meat is added because it has a different prefab name and name all together, so there should be no issues). Note: Fish Meat is bugged. The Raw Version heals for more than its cooked version. Hopefully unintended and to be fixed by Klei within a few days.

Fixes:
Magic Water is officially part of Hamlet (it was commented out until now).
Corrected modes for Nightmare Fuel, Bone Shards, and Volt Goat Horns.
Grammar for Feijodora notes.

Issues:
The DST logic continues to become complicated:

California Roll, Seafood Gumbo, and Surf N Turf are a part of DST now.
	California Roll uses Kelp instead of Seaweed for its recipe. The other two have the same recipe.

The DS Fish (and its cooked variant) no longer exist in-game. 
	It's replaced by Freshwater Fish (which cooks into SW fish), which can be placed in the crockpot and has a .5 fish and .5 meat value (currently commented out).
Two of the new fishes (Corn Cod & Popperfish) might become eligible for fish tacos (its a fish that is also corn, in typical Klei fashion).