6lancmange patch 1 (#5)
* Update recipes.js

Added recipes from Waterlogged (4 with figs), The Curse of the Moon Quay (2 with banana), An Eye for an Eye (bunny stew)

* Update food.js

Added Figs from Waterlogged update, edited banana so it matches the game files (added "_" between cave and banana)

* Update recipes.js

Edited Banana Pop so it works with cave_banana with the underscore. Added DST recipe, although commented out 'cause issue #32

* Update food.js

Added and commented out Crock Pot values for Batilisk Wing – it is only cookable in DST. How to resolve that? Is this issue #32 again?