food.js: _dst and together mode
Added the vanilla + RoG food items to DST.

Added the suffix _dst to all names of foods for DST for both differentiation and as a proper naming scheme. This is needed so that the food item shows up with the right image and correct food values.

Added the together mode for foods for DST.

Added spoiled fish and spoiled fish morsel to DST and added it where needed. Removed the rot of from the ocean fish as well since it was not needed.

Added pondfish, Fish Morsel and Fish Meat Morsel to DST.

This patch will make recipes NOT show up when using the simulator, but will be fixed once recipes.js has been fixed.