Fix Warly multiplier being added twice, added stats to milkmade hat
Fixed a bug where Warly's food multiplier was being applied twice in Warly + Hamlet mode (this does mean that if the user manually selects only Hamlet, the multiplier won't apply, but that is a non-issue imo since SW recipes also work in Hamlet and the user would be missing out on those recipes).

Added the full stats to the milkmade hat.